removed tracked variable.cc
