+++ ngraph_bridge/CMakeLists.txt
@@ -55,7 +55,6 @@ set(SRC
    ngraph_register_stub_kernels.cc   
    ngraph_rewrite_pass.cc
    ngraph_tensor_manager.cc
-   ngraph_tracked_variable.cc
    ngraph_var.cc
    ngraph_utils.cc
    tf_graphcycles.cc
@@ -79,9 +78,6 @@ if(NGRAPH_TF_ENABLE_VARIABLES_AND_OPTIMIZERS)
     list(REMOVE_ITEM SRC ops/ngraph_ops.cc)
     list(APPEND SRC enable_variable_ops/ops/ngraph_ops.cc)
 
-    list(REMOVE_ITEM SRC ngraph_tracked_variable.cc)
-    list(APPEND SRC enable_variable_ops/ngraph_tracked_variable.cc)
-
     # new files
     list(APPEND SRC enable_variable_ops/ngraph_assign_op.cc)
     list(APPEND SRC enable_variable_ops/ngraph_enter_in_catalog.cc)
@@ -91,6 +87,7 @@ if(NGRAPH_TF_ENABLE_VARIABLES_AND_OPTIMIZERS)
     list(APPEND SRC enable_variable_ops/ngraph_rewrite_for_variable_sync.cc)
     list(APPEND SRC enable_variable_ops/ngraph_variable_modifiers.cc)
     list(APPEND SRC enable_variable_ops/ngraph_variable_update_ng_tensor_op.cc)
+    list(APPEND SRC enable_variable_ops/ngraph_variable_op.cc)
 
     add_definitions(-DNGRAPH_TF_ENABLE_VARIABLES_AND_OPTIMIZERS)
 endif()
