--- ngraph_bridge/enable_variable_ops/ngraph_tracked_variable.cc
@@ -1,202 +1,0 @@
-/*******************************************************************************
- * Copyright 2017-2020 Intel Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *******************************************************************************/
-
-#include "tensorflow/core/common_runtime/dma_helper.h"
-#include "tensorflow/core/framework/op.h"
-#include "tensorflow/core/framework/resource_mgr.h"
-#include "tensorflow/core/lib/strings/strcat.h"
-#include "tensorflow/core/platform/default/logging.h"
-
-#include "ngraph/runtime/backend.hpp"
-
-#include "ngraph_bridge/ngraph_backend_manager.h"
-#include "ngraph_bridge/ngraph_catalog.h"
-#include "ngraph_bridge/ngraph_utils.h"
-#include "ngraph_bridge/ngraph_var.h"
-
-using namespace std;
-namespace ng = ngraph;
-
-namespace tensorflow {
-
-namespace ngraph_bridge {
-
-//
-// Forked from tensorflow:tensorflow/core/kernels/variable_ops.{cc,h}
-// and tensorflow:tensorflow/core/ops/state_ops.cc.
-//
-
-// Resource stored by variables in the resource manager
-// (legacy, ref-style version).
-//
-// (Changes: Renamed from LegacyVar, modified to take a TensorShape in
-// constructor.)
-
-/* -------------------------------------------------
-//
-// NGraphVariableOp
-//
----------------------------------------------------*/
-class NGraphVariableOp : public OpKernel {
- public:
-  explicit NGraphVariableOp(OpKernelConstruction* context);
-  ~NGraphVariableOp() override;
-  void Compute(OpKernelContext* ctx) override;
-
- private:
-  int ng_graph_id_;
-  TensorShape shape_;
-  bool update_tf_tensor_;
-  DataType dtype_;
-  string ng_backend_name_;
-  mutex init_mu_;
-  ContainerInfo cinfo_ GUARDED_BY(init_mu_);
-  bool initialized_ GUARDED_BY(init_mu_){false};
-  static int s_instance_count;
-  int my_instance_id{0};
-
-  TF_DISALLOW_COPY_AND_ASSIGN(NGraphVariableOp);
-};
-
-int NGraphVariableOp::s_instance_count = 0;
-
-NGraphVariableOp::NGraphVariableOp(OpKernelConstruction* context)
-    : OpKernel(context),
-      update_tf_tensor_(false),
-      dtype_(RemoveRefType(context->output_type(0))) {
-  my_instance_id = s_instance_count;
-  s_instance_count++;
-
-  OP_REQUIRES_OK(context, context->GetAttr("shape", &shape_));
-  OP_REQUIRES_OK(context,
-                 context->GetAttr("update_tf_tensor", &update_tf_tensor_));
-  OP_REQUIRES_OK(context, context->GetAttr("ngraph_graph_id", &ng_graph_id_));
-  OP_REQUIRES_OK(context,
-                 context->GetAttr("_ngraph_backend", &ng_backend_name_));
-  NGRAPH_VLOG(4) << "NGraphVariable:: Constructor called for: " << def().name()
-                 << " ,update_tf_tensor " << update_tf_tensor_ << " ,Graph ID "
-                 << ng_graph_id_ << " ,backend_name " << ng_backend_name_;
-}
-
-NGraphVariableOp::~NGraphVariableOp() {
-  NGRAPH_VLOG(4) << "~NGraphVariableOp:: " << name() << endl;
-  string node_key = NGraphCatalog::CreateNodeKey(ng_graph_id_, name(), 0);
-  NGraphCatalog::DeleteFromInputVariableSharedNameMap(node_key);
-}
-
-// (Changes: Renamed from VariableOp, modified to pass TensorShape to NGraphVar
-// constructor.)
-void NGraphVariableOp::Compute(OpKernelContext* ctx) {
-  NGRAPH_VLOG(4) << "NGraphVariable:: Compute called for: " << def().name()
-                 << " ,update_tf_tensor " << update_tf_tensor_ << " ,Graph ID "
-                 << ng_graph_id_ << " ,backend_name " << ng_backend_name_;
-
-  std::ostringstream oss;
-  oss << "NGVariable::Compute::" << name();
-  NG_TRACE(oss.str(), name(), "");
-
-  bool log_copies = false;
-  OP_REQUIRES_OK(ctx,
-                 IsNgraphTFLogTensorCopiesEnabled(ng_graph_id_, log_copies));
-  std::stringstream copy_log_str;
-  copy_log_str << "KERNEL[" << type_string() << "]: " << name()
-               << " ,update_tf_tensor " << PrintBool(update_tf_tensor_) << "\n";
-  int number_of_copies = 0;
-
-  mutex_lock l(init_mu_);
-  if (!initialized_) {
-    // Analyze the node attribute of 'ndef' and decides the container and
-    // resource name the kernel should use for accessing the shared
-    // resource.
-    //
-    // 'ndef' is expected to have node attribute "container" and
-    // "shared_name". Returns non-OK if they are not provided or they are
-    // invalid.
-    //
-    // The policy is as following:
-    // * If the attribute "container" is non-empty, it is used as is.
-    //   Otherwise, uses the resource manager's default container.
-    // * If the attribute "shared_name" is non-empty, it is used as is.
-    //   Otherwise, if "use_node_name_as_default" is true, the kernel's
-    //   node name is used as the resource name. Otherwise, a string
-    //   unique to this process is used.
-
-    // API: Status Init(ResourceMgr* rmgr, const NodeDef& ndef,
-    //          bool use_node_name_as_default);
-    //
-    //
-    // We Use context's resource manager's default container
-    // And shared name is same as node_name
-    OP_REQUIRES_OK(ctx, cinfo_.Init(ctx->resource_manager(), def(),
-                                    true /* use name() */));
-    initialized_ = true;
-  }
-
-  auto creator = [this](NGraphVar** var) {
-    NGRAPH_VLOG(4) << "Create NGraphVar Tensors for " << name() << endl;
-    *var = new NGraphVar(dtype_, shape_, ng_backend_name_);
-    return Status::OK();
-  };
-
-  // If "container" has a resource "name", returns it in
-  // "*resource". Otherwise, invokes creator() to create the resource.
-  // The caller takes the ownership of one ref on "*resource".
-  //
-
-  // Here uses the Resource Manager's default container
-  NGraphVar* var;
-  OP_REQUIRES_OK(ctx, cinfo_.resource_manager()->LookupOrCreate<NGraphVar>(
-                          cinfo_.container(), cinfo_.name(), &var, creator));
-
-  // If TF needs the tensor
-  if (update_tf_tensor_) {
-    if (var->copy_ng_to_tf()) {
-      number_of_copies++;
-      copy_log_str << " UPDATE_TF_TENSOR ";
-    }
-    NGRAPH_VLOG(4) << "Copying to TF Tensor";
-  }
-
-  copy_log_str << " Number of copies " << number_of_copies << "\n";
-  if (log_copies) {
-    cout << copy_log_str.str();
-  }
-
-  // Output a reference to our tensor, so it may be updated.
-  //
-  // As long as the resource manager hasn't been cleared the ref we return
-  // here is valid because it owns a ref on var.
-
-  // To output a reference.  Caller retains ownership of mu and tensor_for_ref,
-  // and they must outlive all uses within the step. See comment above.
-  // REQUIRES: IsRefType(expected_output_dtype(index))
-  ctx->set_output_ref(0, var->mu(), var->tensor());
-
-  if (ctx->track_allocations() && var->tensor()->IsInitialized()) {
-    AllocatorAttributes attr;
-    attr.set_gpu_compatible(true);
-    attr.set_nic_compatible(true);
-    ctx->record_persistent_memory_allocation(var->tensor()->AllocatedBytes());
-  }
-  var->Unref();
-}
-
-REGISTER_KERNEL_BUILDER(Name("NGraphVariable").Device(DEVICE_CPU),
-                        NGraphVariableOp);
-
-}  // namespace ngraph_bridge
-
-}  // namespace tensorflow
